refactor: Replace WorkManager alarm scheduling process to AlarmManager #37
