+++ app/src/main/AndroidManifest.xml
@@ -23,12 +23,7 @@
         tools:targetApi="31">
 
         <receiver
-            android:name=".presentation.setting.alarm.AlarmReceiver"
-            android:enabled="true"
-            android:exported="true" />
-
-        <receiver
-            android:name=".receiver.BootCompletedReceiver"
+            android:name=".receiver.DailyAlertBroadcastReceiver"
             android:enabled="true"
             android:exported="true">
             <intent-filter>
@@ -50,11 +45,6 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-
-        <receiver
-            android:name=".receiver.DoneBroadcastReceiver"
-            android:enabled="true"
-            android:exported="false" />
     </application>
 
 </manifest>
