+++ app/src/main/java/com/dev/briefing/di/ManagerUtilModule.kt
@@ -1,6 +1,6 @@
 package com.dev.briefing.di
 
-import com.dev.briefing.util.DailyAlertManager
+import com.dev.briefing.util.dailyalert.DailyAlertManager
 import org.koin.android.ext.koin.androidContext
 import org.koin.dsl.module
